requirements.txt
# Core app
streamlit==1.47.1
python-dotenv==1.1.1
rich==13.7.0
numpy>=1.26,<3.0
pillow>=10.0,<12.0
pydantic>=2.7,<3.0

# OCR + document parsing
easyocr==1.7.1
pdf2image==1.17.0
python-pptx==0.6.23

# Database
psycopg[binary]>=3.2,<4.0
psycopg-pool>=3.2,<4.0

# LangChain + LangGraph (aligned for langchain-core 0.3.x)
langchain>=0.3.27,<0.4.0
langchain-core>=0.3.72,<0.4.0
langchain-community>=0.3.27,<0.4.0
langchain-text-splitters>=0.3.9,<0.4.0
langchain-huggingface>=0.1.2,<0.2.0
langchain-groq>=0.2.0,<0.4.0
langsmith>=0.4.8,<0.5.0
langgraph>=0.2.53,<0.7.0
langgraph-checkpoint>=2.1.2,<3.0.0
langgraph-checkpoint-postgres>=2.0.0,<3.0.0

# Embedding/model stack
sentence-transformers>=3.0,<6.0
transformers>=4.40,<6.0

# Vector database
qdrant-client==1.17.0